Separate Variable class into own file
Variable class is now in gvfx/variables.py and imports are updated accordingly.
